requirements.txt
numpy==1.26.4
scikit-surprise==1.1.3
toolz==0.12.1
